+++ collector.py
@@ -6,6 +6,7 @@ from collectorFilter import UrlCallBack
 from collectorFilter import ContentCallback
 from threadpool import ThreadPool
 from bs4 import BeautifulSoup
+import threading
 
 """
 网页收集器（不支持js执行）
@@ -13,6 +14,8 @@ from bs4 import BeautifulSoup
 
 
 class Collector:
+
+    __local = threading.local()
 
     def __init__(self):
         self.__header = dict()
@@ -61,13 +64,14 @@ class Collector:
                 continue
         return content_str, flag
 
-    def __get_page(self, url, curr_deep):
+    def __get_page(self, url, curr_deep, extend=None):
         """
         获取页面，关键函数，使用深度优先搜索
         此函数中会调用Url黑边名单过滤器、Url回调器、Content回调器
         Args:
             url: 要获取的url
             curr_deep: 当前深度
+            extend: 附加数据
         """
         if curr_deep >= self.__max_deep:
             return
@@ -75,8 +79,9 @@ class Collector:
             return
         self.__visited_url.add(url)
         try:
-            http_client = httplib2.Http('.cache')
-            response, content = http_client.request(url, 'GET', headers=self.__header)
+            if not hasattr(self.__local, "http_client"):
+                self.__local.http_client = httplib2.Http('.cache')
+            response, content = self.__local.http_client.request(url, 'GET', headers=self.__header)
         except Exception as e:
             print("http error", e)
             return
@@ -87,7 +92,7 @@ class Collector:
                 return
             bs = BeautifulSoup(content_str, 'html.parser')
             for k in self.__text_callback:
-                k.callback(url, content_str, bs.title.string)
+                k.callback(url, content_str, bs.title.string, extend)
             link_list = bs.select('a')
             for i in link_list:
                 if i.has_attr('href'):
@@ -95,15 +100,15 @@ class Collector:
                 if i.has_attr('src'):
                     url = i.attrs['src']
                 for k in self.__url_callback:
-                    k.callback(url)
+                    k.callback(url, extend)
                 tmp_url = self.__url_filter.filter(url)
                 if tmp_url is not None:
-                    self.__thread_pool.add_task(self.__get_page, tmp_url, curr_deep + 1)
+                    self.__thread_pool.add_task(self.__get_page, tmp_url, curr_deep + 1, extend)
         else:
             print('status error:', response)
             print(content)
 
-    def start(self, begin_page, deep=1, thread_count=4, wait_exit=True):
+    def start(self, begin_page, deep=1, thread_count=4, wait_exit=True, extend=None):
         """
         开始函数，用于驱动各模块的运行
         Args:
@@ -111,27 +116,33 @@ class Collector:
             deep: 递归深度
             thread_count: 线程数量
             wait_exit: 是否等待所有结果退出
+            extend: 附加数据
         """
         self.__max_deep = deep
         self.__thread_pool.set_work_thread_count(thread_count)
         if isinstance(begin_page, list):
             for page in begin_page:
-                self.__thread_pool.add_task(self.__get_page, page, 0)
+                self.__thread_pool.add_task(self.__get_page, page, 0, extend)
         else:
-            self.__thread_pool.add_task(self.__get_page, begin_page, 0)
+            self.__thread_pool.add_task(self.__get_page, begin_page, 0, extend)
         self.__thread_pool.start()
         if wait_exit:
             self.__thread_pool.exit_when_no_task()
             self.__thread_pool.wait_exit()
 
-    def add_task(self, page, start_deep=0):
+    def add_task(self, page, start_deep=0, extend=None):
         """
         增加新的采集任务
         Args:
-            page: 要采集的页面
+            page: 要采集的页面(集合)
             start_deep: 起始深度
+            extend: 附加数据
         """
-        self.__thread_pool.add_task(self.__get_page, page, start_deep)
+        if isinstance(page, list):
+            for url in page:
+                self.__thread_pool.add_task(self.__get_page, url, start_deep, extend)
+        else:
+            self.__thread_pool.add_task(self.__get_page, page, start_deep, extend)
 
     def set_url_filter(self, url_filter):
         """